% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareEmb}
\alias{compareEmb}
\title{Compares a list of DR embeddings and returns stability statistics}
\usage{
compareEmb(emb_list)
}
\arguments{
\item{emb_list}{A list of embeddings created by createEmb}
}
\value{
Mean of mean correlation, density of mean correlation per embedding, range of mean correlation per embedding, 95\% CI of mean correlation per embedding
}
\description{
Compares a list of input embeddings and aligns them pairwise using procrustes, then computes a Kendall's Tau correlation between each pairwise alignment. Returns the mean of means correlation, density of mean correlations, range of correlations and 95\% CI of means.
}
