% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createEmb}
\alias{createEmb}
\title{Creates a set of n_runs embeddings using either BH t-SNE or UMAP}
\usage{
createEmb(
  dr_input,
  n_runs = 100,
  method = c("umap", "tsne"),
  n_neighbors = 15,
  min_dist = 0.1,
  perplexity = 30,
  theta = 0.5
)
}
\arguments{
\item{dr_input}{Numeric matrix or data frame for DR (rows = observations)}

\item{n_runs}{Integer >= 1; number of embeddings to create}

\item{method}{"umap" or "tsne"}

\item{n_neighbors}{Positive numeric (UMAP)}

\item{min_dist}{Non-negative numeric (UMAP)}

\item{perplexity}{Positive numeric (t-SNE)}

\item{theta}{numeric value between (0,1) (t-SNE)}
}
\value{
A list of length <= n_runs containing DR embeddings (each a matrix).
}
\description{
Using either BH t-SNE or UMAP, creates a list of n_runs embeddings and returns the set of embeddings as a list of dataframes
}
